backend/ml/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Machine Learning libraries
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3

# Geospatial libraries
geopy==2.4.1
shapely==2.0.5

# HTTP client for integration
httpx==0.27.2

# Caching
cachetools==5.4.0

# Model persistence
joblib==1.4.2

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1

# Environment management
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
